calc winner and lift state up
